pyproject.toml
[project]
name = "vis2mobile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "async-lru>=2.0.5",
    "google-genai>=1.56.0",
    "openai>=2.14.0",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "notebook>=7.5.1",
    "poethepoet>=0.38.0",
    "ruff>=0.14.10",
]

[tool.ruff]
extend-exclude = [".venv", ".vscode", ".next", "public", "node_modules"]
